requirements.txt
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
openai==1.30.1
faiss-cpu==1.8.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
tiktoken==0.7.0
numpy==1.26.4
pydantic==2.7.1
httpx==0.27.0
jinja2==3.1.4
python-multipart==0.0.9
